id's are moved to long size messaging structure is optimized md5 is removed crc is added job structure is optimized test code is added log debug level is added

+++ src/Application.cpp
@@ -42,7 +42,7 @@ enum APPTYPE App::getAppType() {
     return type;
 }
 
-bool App::notifyCB(void *context, int target, int id) {
+bool App::notifyCB(void *context, long target, long id) {
 
     return ((App*) context)->notifyHandler(target, id);
 }
